cart

+++ src/components/CartItem.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { useDispatch } from 'react-redux'
-import { addItem, removeItem } from '../Redux/slices/cartSlice'
+import { addItem, minusItem, removeItem } from '../Redux/slices/cartSlice'
 
-function CartItem({ id, title, type, price, count, imageUrl }) {
+function CartItem({ id, title, type, size, price, count, imageUrl }) {
 	const dispatch = useDispatch()
 
 	const onClickPlus = () => {
@@ -10,7 +10,13 @@ function CartItem({ id, title, type, price, count, imageUrl }) {
 	}
 
 	const onClickMinus = () => {
-		dispatch(removeItem(id))
+		dispatch(minusItem(id))
+	}
+
+	const onClickRemove = () => {
+		if (window.confirm('Are you seru you want to remove?')) {
+			dispatch(removeItem(id))
+		}
 	}
 
 	return (
@@ -20,10 +26,14 @@ function CartItem({ id, title, type, price, count, imageUrl }) {
 			</div>
 			<div className='cart__item-info'>
 				<h3>{title}</h3>
-				<p>{type}, 26 см.</p>
+				<p>
+					{type}, {size} см.
+				</p>
 			</div>
 			<div className='cart__item-count'>
-				<div onClick={onClickMinus} className='button button--outline button--circle cart__item-count-minus'>
+				<button
+					onClick={onClickMinus}
+					className='button button--outline button--circle cart__item-count-minus'>
 					<svg
 						width='10'
 						height='10'
@@ -39,9 +49,11 @@ function CartItem({ id, title, type, price, count, imageUrl }) {
 							fill='#EB5A1E'
 						/>
 					</svg>
-				</div>
+				</button>
 				<b>{count}</b>
-				<div onClick={onClickPlus} className='button button--outline button--circle cart__item-count-plus'>
+				<button
+					onClick={onClickPlus}
+					className='button button--outline button--circle cart__item-count-plus'>
 					<svg
 						width='10'
 						height='10'
@@ -57,13 +69,15 @@ function CartItem({ id, title, type, price, count, imageUrl }) {
 							fill='#EB5A1E'
 						/>
 					</svg>
-				</div>
+				</button>
 			</div>
 			<div className='cart__item-price'>
 				<b>{price * count} ₽</b>
 			</div>
 			<div className='cart__item-remove'>
-				<div className='button button--outline button--circle'>
+				<button
+					onClick={onClickRemove}
+					className='button button--outline button--circle'>
 					<svg
 						width='10'
 						height='10'
@@ -79,7 +93,7 @@ function CartItem({ id, title, type, price, count, imageUrl }) {
 							fill='#EB5A1E'
 						/>
 					</svg>
-				</div>
+				</button>
 			</div>
 		</div>
 	)
